Break out modal into it's own component.
